Fix analyze errors: extends vs implements, missing getter, enum names, drift codegen
- All 9 adapters: implements → extends CloudStorageClient (inherits
  default uploadStream/downloadStream/capability getters)
- EncryptedStorageWrapper: add public `inner` getter
- ftp_client_adapter: SecurityType.FTPS → .ftps, FTPEntryType.DIR → .dir
- sync_database.g.dart: add toJson() to data classes, fix uniqueKeys type
- sync_engine/sync_provider: replace Companion.copyWith with direct update
- transfer_provider: fix num→int cast
- connection_dialog: add PanelSide import
- app_state_test: add missing secureStorage param

+++ lib/providers/transfer_provider.dart
@@ -295,7 +295,7 @@ class TransferNotifier extends ChangeNotifier {
           final fileName = p.basename(entity.path);
           final fileData = await localFileService.readFile(entity.path);
           await client.uploadFile(fileData, fileName, newRemotePath);
-          operation.currentBytes += fileData.length;
+          operation.currentBytes += fileData.length as int;
           notifyListeners();
         } catch (_) {}
       } else if (entity is Directory) {
